Split requirement classes
